fixes #6 - The background is gray
Somewhat crude, but I added a method to generate the image art URL according to Ampache API. I create the string directly in AmpacheUtil instead of going through the AmpacheService API. There was no actual way to make the service generate an URL for the method if I was not going to actually execute it, so, well, I just generated the URL and that's it. And given that the MediaPlayerService knows what is the artist ID, I make it send the value in a Bundle back to the Activity, so it can request for the art URL from AmpacheUtil and display it in the background.