re #11 - Adding favorites, featured, etc
Started to add favorites, connecting to the Ampache server using Paging 3 and Rxjava with Flowable! This could be the basis for migrating the other Pagers to that new infrastructure. They additionally need a RemoteMediator to work like they do today, with the database as intermediary. The random albums are listed, but the view is wrong, the albums are listed one below the other