Fixed equality condition
It so happened, that to fix the database reading problems, I had to add a blank Artist when returning the Album from the network, but that caused the equality condition of the album to be ruined, which caused subsequent calls to update the list of songs in the Choose Songs screen to fail. Now that I fixed the equality condition, the Choose Songs screen can be updated correctly.