FirstElement!
It turns out, that I was wrong about the equality condition. I always saw the same list of songs because they were drawn from the MemoryInteractor and the SongsRepository kept only the first element it received. Later, when the network is successfully read, the new songs arrive. This could be fixed by checking if the album to be retrieved from memory is the same that has been requested from memory, or as originally established by the example, if the record expired after some time.